refactor: usage of existing common libs and rename formatters
