+++ src/screens/github-user-projects/columns/column-stats.tsx
@@ -2,7 +2,7 @@ import { EyeOutlined, ForkOutlined, StarOutlined } from "@ant-design/icons";
 import { Tooltip, Typography } from "antd";
 import { createStyles } from "antd-style";
 
-import { formatCount } from "~/utils/number.utils";
+import { formatCount } from "~/utils/number-format.utils";
 
 import type { GithubUserProject } from "../github-user-projects.types";
