+++ src/screens/github-user-projects/github-user-projects-stats.tsx
@@ -3,7 +3,7 @@ import { Card, Col, Row, Statistic, Tag, Typography } from "antd";
 import { createStyles } from "antd-style";
 
 import { useResponsive } from "~/hooks/use-responsive";
-import { formatCount } from "~/utils/number.utils";
+import { formatCount } from "~/utils/number-format.utils";
 
 import type { GithubUserProject } from "./github-user-projects.types";
 import { calculateProjectStats } from "./github-user-projects.utils";
