+++ src/screens/image-ocr/image-ocr.utils.ts
@@ -3,7 +3,7 @@ import { Buffer } from "buffer";
 import { getErrorMessage, isBlank } from "@lichens-innovation/ts-common";
 import Tesseract from "tesseract.js";
 
-import { formatDuration } from "~/utils/number.utils";
+import { formatDuration } from "~/utils/number-format.utils";
 import { countWords } from "~/utils/string.utils";
 
 import type { OcrContext, OcrResult, WorkerStatus } from "./image-ocr.types";
